tests/util: Make purchase_reservation() a separate util #450
